BACKEND ACTUALIZADO/requirements1.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mysql-connector-python==9.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
